Turn f->engine_data into a 64-bit value
Even on 32-bit platforms, we store (uint64_t) -1 in there
for some cases to make it be 'invalid'.

+++ file.h
@@ -97,7 +97,7 @@ struct fio_file {
 	/*
 	 * For use by the io engine
 	 */
-	uintptr_t engine_data;
+	uint64_t engine_data;
 
 	/*
 	 * if io is protected by a semaphore, this is set
